Fix prediction and add Eval

+++ mrcnn/utils.py
@@ -873,7 +873,8 @@ def norm_boxes(boxes, shape):
     h, w = shape
     scale = np.array([h - 1, w - 1, h - 1, w - 1])
     shift = np.array([0, 0, 1, 1])
-    return np.divide((boxes - shift), scale).astype(np.float32)
+    # return np.divide((boxes - shift), scale).astype(np.float32)
+    return np.divide((boxes - shift), scale.astype(np.float32)).astype(np.float32)
 
 
 def denorm_boxes(boxes, shape):
